+++ packages/at_client/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 3.0.51
+- feat: Add atSign to AtSignLoggers' names when relevant, so that log messages are clearer
+- feat: Made notificationService and syncService available via AtClient to enable cleaner simpler code elsewhere
+- fix: Fixed clearing of sync progress listener while switching atsign.
+- fix: Remove the inactive listeners from AtClientManager._changeListeners list.
+- fix: Reverted back path,async packages to older version
 ## 3.0.50
 - feat: changes for at_chops uptake
 - chore: upgrade at_persistence_spec, at_persistence_secondary_server, at_commons version